Fixed a bug in the 1st-person camera movement
There was a bug in the 1st person camera movement when you go far off the platform and then turn around.  The bug occurs when walking forward/backward sometimes where the camera sort of rotates while doing so.  I fixed this by having the modelMatrix translate forward/backward by the lookatdir instead of simply the vector (0, 0, 1).